# Disable HT for DMatrix creation