--- a/src/common/threading_utils.h
+++ b/src/common/threading_utils.h
@@ -160,6 +160,16 @@ inline int32_t OmpSetNumThreads(int32_t* p_threads) {
   omp_set_num_threads(threads);
   return nthread_original;
 }
+inline int32_t OmpSetNumThreadsWithoutHT(int32_t* p_threads) {
+  auto& threads = *p_threads;
+  int32_t nthread_original = omp_get_max_threads();
+  if (threads <= 0) {
+    threads = nthread_original;
+  }
+  omp_set_num_threads(threads);
+  return nthread_original;
+}
+
 }  // namespace common
 }  // namespace xgboost