--- a/src/data/simple_dmatrix.cc
+++ b/src/data/simple_dmatrix.cc
@@ -17,7 +17,6 @@
 #include "../common/random.h"
 #include "../common/threading_utils.h"
 #include "adapter.h"
-
 namespace xgboost {
 namespace data {
 MetaInfo& SimpleDMatrix::Info() { return info_; }
@@ -93,7 +92,7 @@ BatchSet<EllpackPage> SimpleDMatrix::GetEllpackBatches(const BatchParam& param) 
 template <typename AdapterT>
 SimpleDMatrix::SimpleDMatrix(AdapterT* adapter, float missing, int nthread) {
   // Set number of threads but keep old value so we can reset it after
-  int nthread_original = common::OmpSetNumThreads(&nthread);
+  int nthread_original = common::OmpSetNumThreadsWithoutHT(&nthread);
 
   std::vector<uint64_t> qids;
   uint64_t default_max = std::numeric_limits<uint64_t>::max();