Recently introduced `OmpSetNumThreads` function brings up to 2x slowdown on HW with enabled Hyper-threading.

DMatrix create time| master, s | this PR, s | Speedup
-- | -- | -- | --
mortgage | 11.67 | 5.68 | 2.05
higgs | 0.96 | 0.51 | 1.88
msrank | 4.36 | 3.53 | 1.23

It significantly affects inference stage as was mentioned in https://github.com/dmlc/xgboost/issues/6283.

Reason: there is still overheads on memory allocation\initialization (`InitBudget`, `InitStorage`) so there is no gain from usage HT here.