@ShvetsKS 2020-11-13
> So the default setting doesn't account for HT? Also, please don't change the file permission.

@trivialfis by default `nthreads == -1` for DMatrix create, and `omp_get_num_procs` returns 2x of hardware threads with enabled HT. So it has negative impact on time of DMatrix creation always. That's why I reverted initial behavior (before introduction `OmpSetNumThreads` function) for DMatrix.

Files permissions were reverted.