@ShvetsKS 2020-11-13
All tests were passed for previous commit 4320275
Current run failed due to exceeded budget