@ShvetsKS 2020-11-13
@trivialfis would it be better to wait refreshed budget to see all passed tests again?